Share validator between IR + binary-reader-interp (#1346)
The TypeChecker was already shared, but the rest of the other validation
logic was duplicated. This change creates a SharedValidator which is
used by both the Validator and the BinaryReaderInterp classes.

The validator is structured similarly to TypeChecker as a collection of
functions. It's assumed that the functions will be called in the same
order as sections occur in the binary format. The IR valiator does this
too, even though it's possible to validate components out-of-order in
that case.

This change also splits Module validation and Script validation into two
different classes. It should have been written this way in the first
place, and it's nice to do as part of this change since the module
validator logic is mostly moved into the SharedValidator anyway.

Next steps:

* Remove all validation from BinaryReader and move it into the
  SharedValidator.
* Move the TypeChecker into the SharedValidator (maybe not necessary)
* Ensure that validation occurs before creating IR from binary (use
  SharedValidator in BinaryReaderIR? or maybe create
  BinaryReaderValidator passthru that both BinaryReaderIR and
  BinaryReaderInterp use?)
* Rename Validator -> IRValidator, SharedValidator -> Validator