Convert Type from an enum into a class
This is similar to the way Opcode is structured, which allows us to hang
member functions off of the enumeration.

The primary motivator for this change is the GC proposal (and the
function-references proposal) where a Type can be parameterized:

     (type $T (struct ...))
     (func
       (local (ref $T)
       ...
     )

In this case the type is ref, with a parameter of the type index. Making
Type a class will make it easier to store this additional information.