Refactor Const struct's internal storage (#1356)
Const previously stored each value as a union of bit patterns (uint32_t,
uint64_t, v128, etc). It was then extended to support cases where NaN
value (either arithmetic or canonical) was expected.

    bool is_expected_nan;
    union {
      uint32_t u32;
      uint32_t f32_bits;
      ...
      ExpectedNan expected;
    }

With the SIMD proposal, it's possible for each lane of a f32x4 or f64x2
to be a float or an expected NaN, so this doesn't work anymore. It's
possible to move ExpectedNan out of the union, but it's a bit clumsy to
use properly:

    bool is_expected_nan[4];
    ExpectedNan expected[4];
    union { ... }

Instead, I took this as an opportunity to clean up the class a bit.
First, ExpectedNan is extended to handle the case where it is not a NaN
(i.e. not a not a number), which allows us to remove the bool. Then I
store the rest of the data as an array of `uint32_t`, and provide
accessor functions instead.