Fix ubsan issue
`u16 * u16` has the type `int`. If the values are 65535 and 65535, then
the value can't be stored in an int (`0xfffe0001`), which makes UBSan
complain. We want wrapping behavior anyway, so cast this to a T.

Note that this isn't an issue with the other operations, since those all
fit in an `int`.