Add exnref value type to parser
The exnref type was already supported in the type checker, and other
parts of the code, but there was no way to name the type in the text
format.

This PR also fixes makes binary-reader.cc check for just the
exceptions_enabled flag to enable exnref. The exception-handling
proposal depends on the reference types proposal, but that is now
handled at a higher level, in the `UpdateDependencies` function.

Fixes issue #1388.