[wasm2c] Add rot13 example (#1384)
This example demonstrates how to use imported functions. The `rot13`
program takes each command line argument, and rot13-encodes it.

The exported `rot13` function has no arguments, and instead calls back
into the program (via `fill_buf`) with a buffer to fill in. When the
function finishes it calls `buf_done`.

(rot13.wat is the same as in src/test-interp.cc.)