Pass current Thread to host function callbacks (#1412)
The for experimental WASI implementation I'm working on the WASI
callbacks (implemented as HostFuncs) need to know which module a call is
coming from.  This is because the implicitly operate on the memory of
the calling instance.

This approach seems the match the current convention of passing in the
Thread in `Func::DoCall`.

The purpose here is to allow the HostCall callback to determine from
which instance it is being called.  An alternative approach that I
considered was created different set of WASI HostCall object for
each instance so that instance and its memory we bound to the WASI
functions.