[wasm2wat] Write select type immediate
The main fix is in `wat-writer.cc`, where the type immediate was never
being printed. But I've also included a change to how `select` type
immediates are represented in wabt.

Previously, a bare `select` instruction would be stored with the type
`Type::Any`. This is not a real wasm type, and is primarily used for
type validation. The spec instead considers this form of `select` to
have an empty type immediate, which is closer to the `Type::Void` type.

This commit now uses `Type::Void` (or an empty `TypeVector`) to
represent the bare `select` instruction.