Fix resolving local names w/ empty func signatures
This was originally landed in 6bff9f0, but was incorrect (the test
didn't even do the right thing!)

The problem was that the local index was correctly being updated, but it
happened _after_ the names had already been resolved. To fix it,
`ResolveFuncTypes` and `ResolveNamesModule` needed be called in that
order. There was an erroneous assertion that fired as a result ("This
should only be run after resolving names"), but that's incorrect -- the
code works properly before resolving names too, since
`Module::GetFuncType` can properly look up function types by name.

Thanks to @alexcrichton for pointing this out.