Add --relocatable support for tables (#1549)
We add relocations for table numbers on each place where we reify a
table number (call_indirect, table.get, table.set...), but only if
reference types are enabled.

Also, fix symbol table generation with unnamed definitions, to allow for
relocating references to anonymous functions or tables.

As tests, add variants of the relocations and symbol-tables dump tests,
with and without all features enabled.  Enabling reference types causes
relocs to be emitted.  We also add --details to the relocations dump
tests, so that we can see the target symbols for the relocations.