Fix func.wast spec test
Fix spec test failure in func.wast

The wasm text format allows the use of a function type to be specified
one of three ways:

```
;; 1. As an explicit type index or symbol
(type 0)

;; 2. As params and results
(param i32) (result i64)

;; 3. Both; an explicit type index, params and results
(type 0) (param i32) (result i64)
```

Options 2 and 3 are text format sugar, since the binary format always
encodes these as an explicit type index.

Option 3 needs to check that the type at that index matches the
explicitly given params and results. Normally checking whether an index
is in bounds is a validation step, not a syntax checking step. However,
in this case if the index is out of bounds, then the types cannot be
checked, so it must be a syntax error.