Bump testsuite version and fix tests (#1622)
This is an incremental bump of testsuite, the TOT has SIMD tests
which won't work yet.

The main fix is for a test added to the spec repo that has an invalid
functype in the type section. Based on the spec, the functype should be
a single byte 0x60. The leb encoding only comes into effect with the GC
proposal.