Enable bulk memory by default
This features was finished earlier this year:
https://github.com/WebAssembly/proposals/blob/master/finished-proposals.md

One thing to note is that the version of the spec tests we currently
have in third_party/testsuite doesn't have bulk memory merged yet so
this change disables bulk memory when running some of those tests.  This
will be removed in a followup when we update the testsuite.

Fixes: #1717