Set features of the Store used in wasm-interp
Without this change wasm-interp always runs with default
features.

These features seem to be used in just a single location:
src/interp/interp.cc:  int pass = store.features().bulk_memory_enabled() ? Init : Check;

I noticed this when trying to enable bulk memory but run
the interpreter with `--disable-bulk-memory` and it was
not effecting this line of code.