Avoid writing DataCount section when it is not needed.
Without this change, when bulk memory is enabled, the DataCount is
always written when then removed if it proves that its not needed by any
instruction.

However, if we know there are no data segments we can avoid writing it
altogether. This happens to be more compact but I'm hoping to land this
now in anticipation of enabled bulk memory by default.  Once we do that
we don't want all the `dump` tests to include the writing and truncating
of the DataCount section in cases when there is no data.