Fix crash when function counts mismatch
This can happen if we don't stop on first error, and we get a malformed
module where the func counts don't match.

It's hard to write a test for this, since the kStopOnFirstError is fixed
(not set by command line), but this case is quite easy for fuzzers to
catch.