Initial implementation of extended-const proposal.
The primary changes here are to the interpreter and how it handles
initializer expressions.  With this change we model these are normal
function that we run during module initialization.

I imagine we could optimize this further by creating one long function
and encoding the `global.set`/`memory.init`/`table.init` into the
function itself, but this change seems like a good first step to make
the current tests pass.