Share validation code between constant expressions and function bodies. NFC
Previously we has special cases for initializer expressions (constant
expressions).  This change paves the way for adding support for
extended constant expressions that support a wider range of
instructions.

This change removes twice as many lines as it adds which shows that
this simplification is probably worthwhile even without the pending
extensions.