Show tag names in objdump disassembly
Tag names are not officially part of the extended-name-section proposal
(because it only deals with naming things that are in the spec already).

However, I think its reasonable (and useful) to include these names
under a speculative subsection on the basis that tags can only exist
when exceptions are enabled and that engines should ignore unknown name
types.