Fix function body start/end locations in wasm-validate
For text validation, this means the error is always corretly reported on
he final expression in the function.

For binary validation, this means that we report the byte after the last
instrution in the function as the failure location.  This is in line
with other binary validation reports.  For example, for `type mismatch
in i32.add` we report the validtion error at the byte *after* the add
instruction.