Make copies of incoming expression locations in SharedValidator. NFC
The SharedValidator was assuming it could keep a pointer to the incoming
location of each expression longer than the scope of the function.

This assumption holds true for one of the users of SharedValidator
(src/validator.cc) but I'm working on change that adds better location
tracking to the other one in (src/interp/binary-reader-interp.cc) and in
that case its easier to pass temporary locations into the
SharedValidator that don't outlive the call.