Avoid passing signed char to islpha family of function on windows
On windows these functions assert when passed nagative numbers.

This change is the first step in getting wasm2c running on windows.  For
now the C code can be generated on windows but the generated code cannot
is not yet msvc compatible.

Also, update run-spec-wasm2c.py so that it run on windows, although we
are still skipping all the wasm2c tests by default on windows (for now).
Also, honor the --no-compile options which is how I mananged to test
this change locally.