wasm2c signal handler: accommodate platforms where SIGSTKSZ is not a constant
On glibc 2.34+ with the _GNU_SOURCE macro, SIGSTKSZ is no longer a constant
(https://sourceware.org/pipermail/libc-alpha/2021-August/129718.html).

Convert the "alt stack" to a runtime-allocated array, and add wasm_rt_init()
and wasm_rt_free() endpoints to wasm-rt.h that the embedder must call to
set up and clean up the runtime state. (If the embedder doesn't call these,
OOB and exhaustion will result in an uncaught segfault.)