Continue dumping sections even when one is unknown
This is relevant so that when we are dumping a module that contains
unknown sections that the current wabt tooling does not understand, we
still dump all the sections it understands, skipping (and printing an
error) about the unknown sections.

Before this change, we would print partial information, stopping at
the unknown section, without printing the rest of the sections that
wabt could potentially be able to understand.