wasm2c: pass final (non-SIMD) spec tests
To pass the ref_func.wasm and unreached-valid.wast tests,
wasm2c needs to be willing to continue translating a module
that has failed to validate. (The interpreter seems to handle
this similarly.)

Expose this behavior in wasm2c behind a
`--unsafe-ignore-validation-failure` flag, and add the final two
missing tests.