Raise minimum CMake version to 3.16
The stated minimum, 3.1, was released on Dec 17, 2014 and did not
support a value of 17 for CMAKE_CXX_STANDARD. The first version to do so
was 3.8. In fact, attempting to build with CMake 3.7 fails with errors
like:

CMake Error in CMakeLists.txt:
  Target "wasm-strip" requires the language dialect "CXX17" , but CMake does
  not know the compile flags to use to enable it.

So we might as well take this chance to raise the minimum to something
more recent. I propose 3.16 here because of the following features:

* CMP0077 (v3.13) makes the lives of FetchContent users easier by
  allowing option() to be overridden.
* CMP0082 (v3.14) corrects the semantics of install() rules, which is
  relevant for FetchContent users.
* CMake 3.12 allows the namelink of a shared library on Linux to be
  installed separately (i.e. in a dev package)
* CMake 3.14 integrates the standard install destination variables from
  GNUInstallDirs with the install() command.
* CMake 3.15 introduced the CMAKE_MSVC_RUNTIME_LIBRARY variable for
  controlling the selection of the runtime library from the outside.
* CMake 3.16 gained support for generator expressions in RPATH
  properties, which will make it easier to ship shared libraries.

This version is quite conservative and I would even suggest upgrading
further. Windows and macOS users enjoy frequent updates from Homebrew
and Microsoft Visual Studio. Ubuntu Linux 20.04 LTS ships 3.16 and the
newest LTS ships 3.22. Linux users can always install the latest version
through PIP, even on ARM, PowerPC, and s390x, and even without sudo.

Fixes #1568