Add WABT_INSTALL_RULES option
This option is enabled by default when WABT is the top-level project and
disabled by default when it is not. It is used (and should continue to
be used) to guard uses of the install() command.

This fixes an issue where FetchContent users were forced to install
parts of WABT. This most notably included the wasm-rt-impl library and
its associated headers.