Install wasm2c source files as data
Toolchains using wasm2c depend on the wasm-rt files to build
the generated programs. Using a binary wasm-rt library is not enough
for toolchains that do e.g. cross-compilation, and they need to copy
and use the source files instead. Therefore, package these source
files as data for toolchains to use.

Closes #2057.