Enable features in libwabt.js by default
This applies the same defaults for features in the JavaScript-compiled
version as the native tools. Without this change, enabling only the
"reference_types" feature did not work, as it got turned off due to
the dependency on "bulk_memory", which is counter-intuitive (see
Features::UpdateDependencies()).

Additionally, exposes all the features and their default setting in
JavaScript.