wasm2c: multiple .c outputs
This enables wasm2c to have multiple .c outputs, which allows parallel
compilation of the wasm2c outputs. This is useful when the input WASM module is
big.

wasm2c takes the number of .c outputs as an argument to `-j` (default to
1). If the number is equal to 1, the .c output does not change except for two
new macro declarations and the ordering of declarations and definitions. If the
number is greater than 1, wasm2c outputs change in following ways:

1) wasm2c outputs a [module-name]-impl.h that includes any module-wide
declarations, including:
  * content of `WriteSourceTop()`
  * function type declarations
  * tag types
  * tag declarations
  * function declarations
  * data segments and elem segments declarations

Any static declaration become extern in this header.

2) wasm2c outputs [module-name].i.c for i = [0, ..., number of .c outputs). Any
module-wide material is written to [module-name].0.c, including:
  * function types, tags, data segments, elem segments
  * imports and exports
  * module initialization, instantiation and free

3) For each function, wasm2c picks which [module-name].i.c to go in based on the
SHA-256 hash of the function's name. Since SHA-256 is consistent across runs of
wasm2c, only one .c output changes and need to be recompiled if only one
function changes, or one one function is added or moved.