BinaryReaderIR: increase kMaxNestingDepth to 16,384
Rationale: clang (compiled to Wasm) uses a label nesting depth of >1400,
which is greater than the limit of 1024 created in PR #2169.

In practice, with a debug build, values up to about 40,000 appear
sufficient to keep the stack within an 8 MiB limit. On release builds
it's more frugal and we could probably afford an even bigger limit. But,
"16384 ought to be enough nesting depth for anybody."