wasm2c: multiple .c outputs (#2146)
* wasm2c: multiple .c outputs

This enables wasm2c to have multiple .c outputs, which allows parallel
compilation of the wasm2c outputs. This is useful when the input WASM module is
big.

wasm2c takes the number of .c outputs as an argument to `--num-outputs`
(defaulting to 1). If the number is equal to 1, the .c output does not change
except for two new macro declarations and the ordering of declarations and
definitions. If greater than 1, wasm2c outputs change in the following ways:

1) wasm2c outputs a [module-name]-impl.h that includes any module-wide
declarations, including:
  * content of `WriteSourceTop()`
  * function type declarations
  * tag types
  * tag declarations
  * function declarations
  * data segments and elem segments declarations

Any static declaration become extern in this header.

2) wasm2c outputs [module-name]_i.c for i = [0, ..., number of .c outputs - 1). Any
module-wide material is written to [module-name]_0.c, including:
  * function types, tags, data segments, elem segments
  * imports and exports
  * module initialization, instantiation and free

3) For each function implementation, wasm2c assigns it to one output .c file
by sorting the function names and partitioning into roughly equal buckets.
Alternately, the caller can supply its own assignment function (helpful if it wants
the assignments to be more stable in the face of function insertion or deletion).